+++ src/Config.ts
@@ -5,14 +5,22 @@ interface ConfigInterface {
     defaultColors: {
         success: string,
         error: string
+    },
+    spotify: {
+        clientId: string,
+        clientSecret: string
     }
 }
 
 const defaultConfig = {
-    token: "NULL",
+    token: null,
     defaultColors: {
         success: "#4da2b7",
         error: "#bc2727"
+    },
+    spotify: {
+        clientId: null,
+        clientSecret: null
     }
 }
 
@@ -25,5 +33,17 @@ export default function getConfig(): ConfigInterface {
         fs.writeFileSync('./config/config.json', JSON.stringify(defaultConfig))
         throw "Please go fill out the config file!"
     }
-    return JSON.parse(fs.readFileSync('./config/config.json', 'utf-8'));
+
+    var config = JSON.parse(fs.readFileSync('./config/config.json', 'utf-8'));
+
+    for (var c in defaultConfig) {
+        console.log(config[c])
+        if (config[c] == undefined) {
+            config[c] = null;
+            fs.writeFileSync('./config/config.json', JSON.stringify(config))
+            throw "Config file requires attention! There are new variables that were not there before"
+        }
+    }
+
+    return config;
 }