+++ src/Models/LavaplayerManager.ts
@@ -1,6 +1,7 @@
 import { Client, TextChannel } from "discord.js";
 import { Manager } from "erela.js";
-import Spotify from '@kaname-png/erela.js-spotify';
+import Spotify from 'erela.js-spotify';
+import getConfig from "../Config";
 
 
 // Define some options for the node
@@ -9,6 +10,7 @@ const nodes = [{
     password: "youshallnotpass",
     port: 2333,
 }];
+const spotifyCredentials = getConfig().spotify;
 
 var ErelaManager: Manager = null;
 
@@ -25,7 +27,8 @@ function initErela(client: Client) {
         plugins: [
             // Initiate the plugin and pass the two required options.
             new Spotify({
-                limitTracks: Infinity
+                clientID: spotifyCredentials.clientId,
+                clientSecret: spotifyCredentials.clientSecret
             })
         ]
     });